Drop compatibility with PHP 7.2 and 7.3
