+++ .github/workflows/lint.yml
@@ -37,7 +37,7 @@ jobs:
       matrix:
         # Lint against the highest/lowest supported versions of each PHP major.
         # And also do a run against "nightly" (the current dev version of PHP).
-        php_version: ["7.2", "7.4", "8.0", "8.2", "8.3"]
+        php_version: ["7.4", "8.0", "8.2", "8.3"]
 
     name: "Lint: PHP ${{ matrix.php_version }}"
 
@@ -52,7 +52,7 @@ jobs:
       - name: Install PHP 7.x for generating the vendor_prefixed directory
         uses: shivammathur/setup-php@v2
         with:
-          php-version: 7.2
+          php-version: 7.4
           coverage: none
 
       - name: Install Composer dependencies and generate vendor_prefixed directory
