+++ .github/workflows/test.yml
@@ -47,12 +47,12 @@ jobs:
 
     strategy:
       matrix:
-        php_version: ["7.4", "8.0", "8.2"]
+        php_version: ["8.0", "8.2"]
         coverage: [false]
 
         # Run code coverage only on high/low PHP.
         include:
-        - php_version: 7.2
+        - php_version: 7.4
           coverage: true
         - php_version: 8.3
           coverage: true
@@ -67,7 +67,7 @@ jobs:
       - name: Install PHP 7.x for generating the vendor_prefixed directory and dependency injection
         uses: shivammathur/setup-php@v2
         with:
-          php-version: 7.2
+          php-version: 7.4
           coverage: none
 
       - name: Install Composer dependencies, generate vendor_prefixed directory and run dependency injection
@@ -137,12 +137,12 @@ jobs:
     strategy:
       matrix:
         include:
-          - php_version: "7.2"
+          - php_version: "7.4"
             wp_version: "6.6"
             multisite: true
             coverage: true
 
-          - php_version: "7.3"
+          - php_version: "7.4"
             wp_version: "trunk"
             multisite: true
             coverage: false
@@ -179,9 +179,9 @@ jobs:
 
     services:
       mysql:
-        # Use MySQL 5.6 for PHP 7.2, use MySQL 5.7 for PHP 7.3 < 7.4, otherwise MySQL 8.0.
+        # Use MySQL 5.7 for PHP 7.4, otherwise MySQL 8.0.
         # Also see: https://core.trac.wordpress.org/ticket/52496
-        image: mysql:${{ ( matrix.php_version == '7.2' && '5.6' ) || ( matrix.php_version < '7.4' && '5.7' ) || '8.0' }}
+        image: mysql:${{ ( matrix.php_version == '7.4' && '5.7' ) || '8.0' }}
         env:
           MYSQL_ALLOW_EMPTY_PASSWORD: false
         ports:
@@ -199,7 +199,7 @@ jobs:
       - name: Install PHP 7.x for generating the vendor_prefixed directory and dependency injection
         uses: shivammathur/setup-php@v2
         with:
-          php-version: 7.2
+          php-version: 7.4
           coverage: none
 
       - name: Install Composer dependencies, generate vendor_prefixed directory and run dependency injection
