+++ tests/Unit/bootstrap.php
@@ -39,7 +39,7 @@ if ( ! \defined( 'WPSEO_BASENAME' ) ) {
 \define( 'YOAST_VENDOR_DEFINE_PREFIX', 'YOASTSEO_VENDOR__' );
 \define( 'YOAST_VENDOR_PREFIX_DIRECTORY', 'vendor_prefixed' );
 
-\define( 'YOAST_SEO_PHP_REQUIRED', '7.2.5' );
+\define( 'YOAST_SEO_PHP_REQUIRED', '7.4' );
 \define( 'YOAST_SEO_WP_TESTED', '6.8' );
 \define( 'YOAST_SEO_WP_REQUIRED', '6.6' );
